stbt batch report: Use fixed table-layout
And don't let table cells wrap to more than 1 line; truncate the content
using CSS. This way you can fit more rows in a single screen.

This also prevents the table from ever exceeding the width of its parent
div. If you have lots of extra-columns each column will be quite narrow,
but at least it will be visible (and searchable). They could be so
narrow that I had to add `th { word-wrap: break-word }` so that the
heading is still completely visible.

Since we're using CSS for the table-cell truncation we can include the
full failure-reason and notes in the html, which means that the
client-side search/filtering works on those fields.